Updated Controller

+++ project/com/controller/DatasetController.py
@@ -1,10 +1,11 @@
-from flask import request, render_template, redirect, url_for
-from project import app
-from werkzeug.utils import secure_filename
 import os
+from datetime import date, datetime
+from flask import request, render_template, redirect, url_for
+from werkzeug.utils import secure_filename
+from project import app
+from project.com.controller.LoginController import adminLoginSession
 from project.com.dao.DatasetDAO import DatasetDAO
 from project.com.vo.DatasetVO import DatasetVO
-from datetime import date, datetime
 
 UPLOAD_FOLDER = 'project/static/adminResource/dataset/'
 app.config['UPLOAD_FOLDER'] = UPLOAD_FOLDER
@@ -13,7 +14,10 @@ app.config['UPLOAD_FOLDER'] = UPLOAD_FOLDER
 @app.route('/admin/loadDataset', methods=['GET'])
 def adminLoadDataset():
     try:
-        return render_template('admin/addDataset.html')
+        if adminLoginSession() == 'admin':
+            return render_template('admin/addDataset.html')
+        else:
+            return redirect(url_for('adminLogoutSession'))
     except Exception as ex:
         print(ex)
 
@@ -21,36 +25,39 @@ def adminLoadDataset():
 @app.route('/admin/insertDataset', methods=['POST'])
 def adminInsertDataset():
     try:
-        app.config['UPLOAD_FOLDER'] = UPLOAD_FOLDER
+        if adminLoginSession() == 'admin':
+            app.config['UPLOAD_FOLDER'] = UPLOAD_FOLDER
 
-        datasetVO = DatasetVO()
-        datasetDAO = DatasetDAO()
+            datasetVO = DatasetVO()
+            datasetDAO = DatasetDAO()
 
-        file = request.files['datasetFileName']
-        print(file)
+            file = request.files['datasetFileName']
+            print(file)
 
-        datasetFilename = secure_filename(file.filename)
-        print(datasetFilename)
+            datasetFilename = secure_filename(file.filename)
+            print(datasetFilename)
 
-        datasetFilepath = os.path.join(app.config['UPLOAD_FOLDER'])
-        print(datasetFilepath)
+            datasetFilepath = os.path.join(app.config['UPLOAD_FOLDER'])
+            print(datasetFilepath)
 
-        file.save(os.path.join(datasetFilepath, datasetFilename))
+            file.save(os.path.join(datasetFilepath, datasetFilename))
 
-        datasetUploaddate = date.today()
-        print(datasetUploaddate)
+            datasetUploaddate = date.today()
+            print(datasetUploaddate)
 
-        datasetUploadtime = datetime.now().strftime("%H:%M:%S")
-        print(datasetUploadtime)
+            datasetUploadtime = datetime.now().strftime("%H:%M:%S")
+            print(datasetUploadtime)
 
-        datasetVO.datasetFilename = datasetFilename
-        datasetVO.datasetFilepath = datasetFilepath.replace('project', '..')
-        datasetVO.datasetUploaddate = datasetUploaddate
-        datasetVO.datasetUploadtime = datasetUploadtime
+            datasetVO.datasetFilename = datasetFilename
+            datasetVO.datasetFilepath = datasetFilepath.replace('project', '..')
+            datasetVO.datasetUploaddate = datasetUploaddate
+            datasetVO.datasetUploadtime = datasetUploadtime
 
-        datasetDAO.insertDataset(datasetVO)
+            datasetDAO.insertDataset(datasetVO)
 
-        return redirect(url_for('adminSearchDataset'))
+            return redirect(url_for('adminSearchDataset'))
+        else:
+            return redirect(url_for('adminLogoutSession'))
     except Exception as ex:
         print(ex)
 
@@ -58,9 +65,12 @@ def adminInsertDataset():
 @app.route('/admin/searchDataset', methods=['GET'])
 def adminSearchDataset():
     try:
-        datasetDAO = DatasetDAO()
-        datasetVOList = datasetDAO.viewDataset()
-        return render_template('admin/viewDataset.html', datasetVOList=datasetVOList)
+        if adminLoginSession() == 'admin':
+            datasetDAO = DatasetDAO()
+            datasetVOList = datasetDAO.viewDataset()
+            return render_template('admin/viewDataset.html', datasetVOList=datasetVOList)
+        else:
+            return redirect(url_for('adminLogoutSession'))
     except Exception as ex:
         print(ex)
 
@@ -68,15 +78,18 @@ def adminSearchDataset():
 @app.route('/admin/deleteDataset', methods=['GET'])
 def adminDeleteDataset():
     try:
-        datasetVO = DatasetVO()
-        datasetDAO = DatasetDAO()
+        if adminLoginSession() == 'admin':
+            datasetVO = DatasetVO()
+            datasetDAO = DatasetDAO()
 
-        datasetId = request.args.get('datasetId')
+            datasetId = request.args.get('datasetId')
 
-        datasetVO.datasetId = datasetId
+            datasetVO.datasetId = datasetId
 
-        datasetDAO.deleteDataset(datasetVO)
+            datasetDAO.deleteDataset(datasetVO)
 
-        return redirect(url_for('adminViewDataset'))
+            return redirect(url_for('adminSearchDataset'))
+        else:
+            return redirect(url_for('adminLogoutSession'))
     except Exception as ex:
         print(ex)
