+++ project/com/controller/ZoneController.py
@@ -1,18 +1,18 @@
 from flask import request, render_template, redirect, url_for
+
 from project import app
+from project.com.controller.LoginController import adminLoginSession
 from project.com.dao.ZoneDAO import ZoneDAO
 from project.com.vo.ZoneVO import ZoneVO
-
-
-# @app.route('/')
-# def adminLoadDashboard():
-#     return render_template('admin/index.html')
 
 
 @app.route('/admin/loadZone', methods=['GET'])
 def adminLoadZone():
     try:
-        return render_template('admin/addZone.html')
+        if adminLoginSession() == 'admin':
+            return render_template('admin/addZone.html')
+        else:
+            return redirect(url_for('adminLoadDashboard'))
     except Exception as ex:
         print(ex)
 
@@ -20,18 +20,21 @@ def adminLoadZone():
 @app.route('/admin/insertZone', methods=['POST'])
 def adminInsertZone():
     try:
-        zoneName = request.form['zoneName']
-        zoneDescription = request.form['zoneDescription']
+        if adminLoginSession() == 'admin':
+            zoneName = request.form['zoneName']
+            zoneDescription = request.form['zoneDescription']
 
-        zoneVO = ZoneVO()
-        zoneDAO = ZoneDAO()
+            zoneVO = ZoneVO()
+            zoneDAO = ZoneDAO()
 
-        zoneVO.zoneName = zoneName
-        zoneVO.zoneDescription = zoneDescription
+            zoneVO.zoneName = zoneName
+            zoneVO.zoneDescription = zoneDescription
 
-        zoneDAO.insertZone(zoneVO)
+            zoneDAO.insertZone(zoneVO)
 
-        return redirect(url_for('adminSearchZone'))
+            return redirect(url_for('adminSearchZone'))
+        else:
+            return redirect(url_for('adminLoadDashboard'))
     except Exception as ex:
         print(ex)
 
@@ -39,10 +42,13 @@ def adminInsertZone():
 @app.route('/admin/searchZone', methods=['GET'])
 def adminSearchZone():
     try:
-        zoneDAO = ZoneDAO()
-        zoneVOList = zoneDAO.viewZone()
-        print("__________________", zoneVOList)
-        return render_template('admin/viewZone.html', zoneVOList=zoneVOList)
+        if adminLoginSession() == 'admin':
+            zoneDAO = ZoneDAO()
+            zoneVOList = zoneDAO.viewZone()
+            print("__________________", zoneVOList)
+            return render_template('admin/viewZone.html', zoneVOList=zoneVOList)
+        else:
+            return redirect(url_for('adminLoadDashboard'))
     except Exception as ex:
         print(ex)
 
@@ -50,17 +56,20 @@ def adminSearchZone():
 @app.route('/admin/deleteZone', methods=['GET'])
 def adminDeleteZone():
     try:
-        zoneVO = ZoneVO()
+        if adminLoginSession() == 'admin':
+            zoneVO = ZoneVO()
 
-        zoneDAO = ZoneDAO()
+            zoneDAO = ZoneDAO()
 
-        zoneId = request.args.get('zoneId')
+            zoneId = request.args.get('zoneId')
 
-        zoneVO.zoneId = zoneId
+            zoneVO.zoneId = zoneId
 
-        zoneDAO.deleteZone(zoneVO)
+            zoneDAO.deleteZone(zoneVO)
 
-        return redirect(url_for('adminSearchZone'))
+            return redirect(url_for('adminSearchZone'))
+        else:
+            return redirect(url_for('adminLoadDashboard'))
     except Exception as ex:
         print(ex)
 
@@ -68,21 +77,24 @@ def adminDeleteZone():
 @app.route('/admin/editZone', methods=['GET'])
 def adminEditZone():
     try:
-        zoneVO = ZoneVO()
+        if adminLoginSession() == 'admin':
+            zoneVO = ZoneVO()
 
-        zoneDAO = ZoneDAO()
+            zoneDAO = ZoneDAO()
 
-        zoneId = request.args.get('zoneId')
+            zoneId = request.args.get('zoneId')
 
-        zoneVO.zoneId = zoneId
+            zoneVO.zoneId = zoneId
 
-        zoneVOList = zoneDAO.editZone(zoneVO)
+            zoneVOList = zoneDAO.editZone(zoneVO)
 
-        print("=======zoneVOList=======", zoneVOList)
+            print("=======zoneVOList=======", zoneVOList)
 
-        print("=======type of zoneVOList=======", type(zoneVOList))
+            print("=======type of zoneVOList=======", type(zoneVOList))
 
-        return render_template('admin/editZone.html', zoneVOList=zoneVOList)
+            return render_template('admin/editZone.html', zoneVOList=zoneVOList)
+        else:
+            return redirect(url_for('adminLoadDashboard'))
     except Exception as ex:
         print(ex)
 
@@ -90,19 +102,22 @@ def adminEditZone():
 @app.route('/admin/updateZone', methods=['POST'])
 def adminUpdateZone():
     try:
-        zoneId = request.form['zoneId']
-        zoneName = request.form['zoneName']
-        zoneDescription = request.form['zoneDescription']
+        if adminLoginSession() == 'admin':
+            zoneId = request.form['zoneId']
+            zoneName = request.form['zoneName']
+            zoneDescription = request.form['zoneDescription']
 
-        zoneVO = ZoneVO()
-        zoneDAO = ZoneDAO()
+            zoneVO = ZoneVO()
+            zoneDAO = ZoneDAO()
 
-        zoneVO.zoneId = zoneId
-        zoneVO.zoneName = zoneName
-        zoneVO.zoneDescription = zoneDescription
+            zoneVO.zoneId = zoneId
+            zoneVO.zoneName = zoneName
+            zoneVO.zoneDescription = zoneDescription
 
-        zoneDAO.updateZone(zoneVO)
+            zoneDAO.updateZone(zoneVO)
 
-        return redirect(url_for('adminSearchZone'))
+            return redirect(url_for('adminSearchZone'))
+        else:
+            return redirect(url_for('adminLoadDashboard'))
     except Exception as ex:
         print(ex)
